325: Hide most categories for lower-level users
Fixes humanmade/wikimedia#325: for users who are not an admin or editor,
hide any categories in the block editor category list which do not
occur in the top navigation.

Updating the navigation will flush the cache of which categories can be
selected. Higher-level users will get the full, unfiltered list.